+++ docs/src/examples/components/Popup/Types/PopupExampleInline.shorthand.tsx
@@ -1,0 +1,12 @@
+import * as React from 'react'
+import { Button, Popup } from '@stardust-ui/react'
+
+const PopupExample = () => (
+  <Popup
+    trigger={<Button icon="expand" />}
+    content="This popup is rendered next to the trigger."
+    inline
+  />
+)
+
+export default PopupExample